feat: enrich musea
