+++ npm/musea-mcp-server/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@vizejs/musea-mcp-server",
   "version": "0.0.1-alpha.72",
-  "description": "MCP server for Musea - AI-accessible component gallery",
+  "description": "MCP server for building Vue.js design systems - component analysis, documentation, variant generation, and design tokens",
   "type": "module",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
@@ -30,7 +30,8 @@
     "mcp",
     "model-context-protocol",
     "vue",
-    "component-gallery",
+    "design-system",
+    "component-analysis",
     "musea",
     "ai"
   ],
